+++ noodles-sam/src/alignment/record/data_ref.rs
@@ -3,5 +3,5 @@ use super::Data;
 #[doc(hidden)]
 pub enum DataRef<'a> {
     FieldEncoded(&'a [u8]),
-    Data(Box<dyn Data + 'a>),
+    Data(Box<dyn Data<'a> + 'a>),
 }